missing refresh token error on logout corrected

+++ services/authentication.service.js
@@ -145,6 +145,7 @@ const refreshAccessToken = async (refreshTokenValue) => {
 };
 
 const logOutUser = async (refreshTokenValue) => {
+  if (!refreshTokenValue) return;
   await deleteRefreshToken(refreshTokenValue);
 };
